// prisma/generated/type-graphql/models/Coupons.d.ts
import { Products } from "../models/Products";
export declare class Coupons {
    id: bigint;
    name: string;
    excerpt?: string | null;
    discount?: string | null;
    image?: string | null;
    published?: Date | null;
    expiration?: string | null;
    categories?: string | null;
    articles?: string | null;
    products?: string | null;
    customers?: string | null;
    users?: string | null;
    prod_id: bigint;
    products_couponsToproducts?: Products;
}