// prisma/generated/type-graphql/models/Product_attribute.d.ts
import { Product_attribute_set } from "../models/Product_attribute_set";
import { Products } from "../models/Products";
import { Product_attributeCount } from "../resolvers/outputs/Product_attributeCount";
export declare class Product_attribute {
    id: number;
    default_label?: string | null;
    catalog_input_type_store_owner?: string | null;
    values_required?: boolean | null;
    attribute_code?: string | null;
    scope?: string | null;
    default_value?: string | null;
    unique_value?: boolean | null;
    input_validation_store_owner?: string | null;
    add_column_options?: boolean | null;
    filter_options?: boolean | null;
    use_search?: boolean | null;
    comparable_storefront?: boolean | null;
    layered_navigation?: boolean | null;
    search_results_layered_navigation?: boolean | null;
    position?: string | null;
    promo_rule_conditions?: boolean | null;
    allow_html_tags_storefront?: boolean | null;
    visible_catalog_pages_storefront?: boolean | null;
    used_product_listing?: boolean | null;
    used_sorting_product_listing?: boolean | null;
    prod_id: bigint;
    products?: Products;
    product_attribute_set?: Product_attribute_set[];
    _count?: Product_attributeCount | null;
}