// prisma/generated/type-graphql/models/Product_types.d.ts
export declare class Product_types {
    id: number;
    type_name?: string | null;
    taxes?: string | null;
    isShippable?: string | null;
    meta_url?: string | null;
    meta_description?: string | null;
    filter_options?: string | null;
    product_type?: string | null;
    prod_id: bigint;
    products?: string | null;
}