// prisma/generated/type-graphql/models/Workspaces.d.ts
export declare class Workspaces {
    id: bigint;
    created_at?: Date | null;
    code?: string | null;
    name?: string | null;
    description?: string | null;
    customers?: number | null;
    users?: number | null;
    products?: bigint | null;
    tasks?: bigint | null;
    brands?: bigint | null;
    shops?: number | null;
    category?: number | null;
    isPublic?: boolean | null;
    projects?: bigint | null;
    author?: string | null;
    media?: string | null;
    brands_brandsToworkspaces?: string | null;
    categories?: string | null;
    customers_customersToworkspaces?: string | null;
    products_productsToworkspaces?: string | null;
    projects_projectsToworkspaces?: string | null;
    shops_shopsToworkspaces?: string | null;
    tasks_tasksToworkspaces?: string | null;
    users_usersToworkspaces?: string | null;
}